Fix Float32/Float64 type mismatch in Polars map_elements calls

+++ src/pav3/align/score.py
@@ -138,7 +138,7 @@ class ScoreModel(ABC):
             .select(pl.col('align_ops'))
             .to_series()
             .map_elements(op.row_to_arr, return_dtype=pl.Object)
-            .map_elements(self.score_op_arr, return_dtype=pl.Float32)
+            .map_elements(self.score_op_arr, return_dtype=pl.Float64)
         )
 
         # A pure polars implementation is slower:
